Complete

+++ controllers/contactController.js
@@ -2,16 +2,18 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc get all contacts
 //@route Get /api/contacts
-//@access public
+//@access private
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({
+    user_id: req.user.id,
+  });
   res.status(200).json(contacts);
 });
 
 //@desc Create new contacts
 //@route post /api/contacts
-//@access public
+//@access private
 
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is ", req.body);
@@ -21,13 +23,18 @@ const createContact = asyncHandler(async (req, res) => {
     throw new Error("All fields are required");
   }
 
-  const contact = await Contact.create({ name, email, phone });
+  const contact = await Contact.create({
+    name,
+    email,
+    phone,
+    user_id: req.user.id,
+  });
   res.status(201).json(contact);
 });
 
 //@desc get contact for a single id
 //@route get /api/contacts/:id
-//@access public
+//@access private
 
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -40,13 +47,18 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update contact
 //@route put /api/contacts/:id
-//@access public
+//@access private
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("No contact found");
+  }
+
+  if(contact.user_id.toString() !== req.user.id){
+    res.status(403);
+    throw new Error("User does not have permission to update other user")
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
@@ -59,7 +71,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc delete contact
 //@route delete /api/contacts/:id
-//@access public
+//@access private
 
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
@@ -68,6 +80,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     throw new Error("No contact found");
   }
 
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User does not have permission to delete other user")
+  }
   await Contact.findByIdAndRemove(req.params.id, req.body);
   res.status(200).json(contact);
 });
